# Fix comparison