@dfilatov 2015-11-12
But with my laptop and nodejs 4.2.1 I've got quite different result:
```
               mean time ops/sec
Q              5.282ms   189
When           1.881ms   532
Bluebird       2.137ms   468
Pinkie         2.936ms   341
ES2015 Promise 2.915ms   343
Vow            1.861ms   537
```