@awinogradov 2015-11-12
Wtf this https://github.com/petkaantonov/bluebird/tree/master/benchmark ?